[Fusion] Fold bank tests into infra and drop slide-perf harness
Keep CI coverage on the semaphore bank without a separate env-gated
benchmark suite, and present the demo numbers as a standalone reference.
