+++ tests/ttnn/unit_tests/operations/fused/parallel_sequential/demo/test_fused_demo.py
@@ -30,10 +30,7 @@ To run profiling modes, add the desired values back to the @pytest.mark.parametr
 list in the test, then optionally filter with: pytest ... -k cold_start
 """
 
-import gc
-import os
 import shutil
-import statistics
 import time
 from pathlib import Path
 
@@ -56,12 +53,6 @@ COMPUTE_CONFIG = ttnn.WormholeComputeKernelConfig(
     math_fidelity=ttnn.MathFidelity.HiFi4,
     math_approx_mode=False,
     fp32_dest_acc_en=True,
-)
-
-RUN_SLIDE_PERF = os.environ.get("TT_METAL_RUN_FUSION_SLIDE_PERF", "").strip().lower() in {"1", "true", "yes", "on"}
-SLIDE_PERF_SKIP = pytest.mark.skipif(
-    not RUN_SLIDE_PERF,
-    reason="Manual slide benchmark; set TT_METAL_RUN_FUSION_SLIDE_PERF=1 to run",
 )
 
 
@@ -301,51 +292,6 @@ def _time_e2e(fn, device, num_warmup=5, num_measure=100):
         fn()
     sync()
     return 1000 * (time.perf_counter() - t0) / num_measure
-
-
-def _benchmark_e2e_mode(fn, device, *, num_trials=7, num_measure=100):
-    """Measure one mode in isolation and return its median and raw samples."""
-    for _ in range(5):
-        fn()
-    ttnn.synchronize_device(device)
-
-    samples_us = [_time_e2e(fn, device, num_warmup=0, num_measure=num_measure) * 1000 for _ in range(num_trials)]
-    return statistics.median(samples_us), samples_us
-
-
-def _print_slide_perf_result(demo, mode, median_us, samples_us):
-    print(f"\nFUSION_SLIDE_PERF demo={demo} metric=steady_state_throughput")
-    samples = ",".join(f"{sample:.3f}" for sample in samples_us)
-    print(f"FUSION_SLIDE_PERF demo={demo} mode={mode} " f"median_us={median_us:.3f} samples_us={samples}")
-
-
-def _benchmark_barrier_per_transition(fused_fn, baseline_fn, device, num_transitions, *, num_trials=7, num_measure=100):
-    """Measure barrier cost by subtracting a one-phase dispatch in paired trials."""
-    for fn in (fused_fn, baseline_fn):
-        for _ in range(5):
-            fn()
-        ttnn.synchronize_device(device)
-
-    samples_us = []
-    for trial in range(num_trials):
-        if trial % 2 == 0:
-            fused_ms = _time_e2e(fused_fn, device, num_warmup=0, num_measure=num_measure)
-            baseline_ms = _time_e2e(baseline_fn, device, num_warmup=0, num_measure=num_measure)
-        else:
-            baseline_ms = _time_e2e(baseline_fn, device, num_warmup=0, num_measure=num_measure)
-            fused_ms = _time_e2e(fused_fn, device, num_warmup=0, num_measure=num_measure)
-        samples_us.append((fused_ms - baseline_ms) * 1000 / num_transitions)
-    return statistics.median(samples_us), samples_us
-
-
-def _run_device_profile(run, device, *, num_warmup=1, num_measure=5):
-    """Emit repeated synchronized dispatches for robust Tracy medians."""
-    for _ in range(num_warmup):
-        run()
-        ttnn.synchronize_device(device)
-    for _ in range(num_measure):
-        run()
-        ttnn.synchronize_device(device)
 
 
 # =============================================================================
@@ -1187,19 +1133,11 @@ class TestPerfDemos:
             shards,
         )
 
-    def _sharded_tree_make_ops(self, device, *, setup=None, persistent=False):
-        """Create all OpDescriptors for the sharded tree.
-
-        ``setup`` lets throughput benchmarks reuse the same model tensors while
-        constructing fresh inline descriptors. ``persistent`` omits the root
-        activation so the returned tree can be reused via ``ln_stem.update()``.
-        """
+    def _sharded_tree_make_ops(self, device):
+        """Create all OpDescriptors for the sharded tree."""
         from models.experimental.ops.descriptors.normalization import layer_norm
         from models.experimental.ops.descriptors.data_movement.slice import slice
         from models.experimental.ops.descriptors.matmul import matmul as matmul_desc
-
-        if setup is None:
-            setup = self._sharded_tree_setup(device)
 
         (
             stem_cores,
@@ -1216,15 +1154,16 @@ class TestPerfDemos:
             tt_B_left,
             tt_B_right,
             shards,
-        ) = setup
+        ) = self._sharded_tree_setup(device)
 
         rows, cols = 2048, 256
         half = rows // 2
         quarter = rows // 4
 
         # Level 0: stem LN on 16 cores (auto-detects sharded input grid)
-        stem_kwargs = dict(core_range_set=stem_cores, epsilon=1e-5, compute_kernel_config=COMPUTE_CONFIG)
-        ln_stem = layer_norm.layer_norm(**stem_kwargs) if persistent else layer_norm.layer_norm(tt_input, **stem_kwargs)
+        ln_stem = layer_norm.layer_norm(
+            tt_input, core_range_set=stem_cores, epsilon=1e-5, compute_kernel_config=COMPUTE_CONFIG
+        )
 
         # Level 1: slice (sharded on 8 cores) -> matmul (sharded output on 8 cores)
         sl_top = slice(
@@ -2069,536 +2008,6 @@ class TestPerfDemos:
         else:
             pytest.fail(f"Unsupported perf_mode={perf_mode!r}")
 
-    # -----------------------------------------------------------------
-    # Manual slide benchmarks — comparative demos
-    # -----------------------------------------------------------------
-
-    def _linear_chain_perf_setup(self, device, H):
-        core_range, mm_cfg, torch_input, torch_w, torch_b = self._linear_chain_setup(device, H)
-        dram = ttnn.DRAM_MEMORY_CONFIG
-
-        def to_device(tensor):
-            return ttnn.from_torch(
-                tensor,
-                dtype=ttnn.bfloat16,
-                layout=ttnn.TILE_LAYOUT,
-                device=device,
-                memory_config=dram,
-            )
-
-        compute_cfg = ttnn.WormholeComputeKernelConfig(
-            math_fidelity=ttnn.MathFidelity.HiFi4,
-            math_approx_mode=False,
-            fp32_dest_acc_en=False,
-        )
-        return (
-            core_range,
-            mm_cfg,
-            to_device(torch_input),
-            to_device(torch_w),
-            to_device(torch_w),
-            to_device(torch_b),
-            compute_cfg,
-        )
-
-    def _linear_chain_fused_callable(self, setup, *, persistent):
-        from models.experimental.ops.descriptors.fusion import Sequential
-        from models.experimental.ops.descriptors.matmul import matmul as matmul_desc
-        from models.experimental.ops.descriptors.normalization import rms_norm
-
-        core_range, mm_cfg, tt_input, tt_w1, tt_w2, tt_b, compute_cfg = setup
-
-        def make_container():
-            root_kwargs = dict(
-                core_range_set=core_range,
-                weight=tt_w1,
-                epsilon=1e-5,
-                compute_kernel_config=compute_cfg,
-            )
-            r1 = rms_norm.rms_norm(**root_kwargs) if persistent else rms_norm.rms_norm(tt_input, **root_kwargs)
-            matmul = matmul_desc(
-                r1.output_tensors[0],
-                tt_b,
-                core_range_set=core_range,
-                program_config=mm_cfg,
-                compute_kernel_config=compute_cfg,
-            )
-            r2 = rms_norm.rms_norm(
-                matmul.output_tensors[0],
-                core_range_set=core_range,
-                weight=tt_w2,
-                epsilon=1e-5,
-                compute_kernel_config=compute_cfg,
-            )
-            return Sequential(r1, matmul, r2), r1
-
-        if persistent:
-            container, root = make_container()
-
-            def run_persistent():
-                root.update(tt_input)
-                return container.run()
-
-            return run_persistent
-
-        def run_inline():
-            container, _ = make_container()
-            return container.run()
-
-        return run_inline
-
-    @staticmethod
-    def _linear_chain_unfused_callable(setup):
-        _, mm_cfg, tt_input, tt_w, _, tt_b, _ = setup
-
-        def run_unfused():
-            result = ttnn.rms_norm(tt_input, weight=tt_w, epsilon=1e-5)
-            result = ttnn.matmul(result, tt_b, program_config=mm_cfg)
-            return ttnn.rms_norm(result, weight=tt_w, epsilon=1e-5)
-
-        return run_unfused
-
-    def _sharded_chain_fused_callable(self, setup, *, persistent):
-        from models.experimental.ops.descriptors.fusion import Sequential
-        from models.experimental.ops.descriptors.normalization import layer_norm, rms_norm
-
-        cores, sharded_mem, _, tt_input, tt_w, _, _ = setup
-
-        def make_container():
-            root_kwargs = dict(
-                core_range_set=cores,
-                weight=tt_w,
-                epsilon=1e-5,
-                compute_kernel_config=COMPUTE_CONFIG,
-                memory_config=sharded_mem,
-            )
-            root = rms_norm.rms_norm(**root_kwargs) if persistent else rms_norm.rms_norm(tt_input, **root_kwargs)
-            leaf = layer_norm.layer_norm(
-                root.output_tensors[0],
-                core_range_set=cores,
-                weight=tt_w,
-                epsilon=1e-5,
-                compute_kernel_config=COMPUTE_CONFIG,
-                memory_config=sharded_mem,
-            )
-            return Sequential(root, leaf), root
-
-        if persistent:
-            container, root = make_container()
-
-            def run_persistent():
-                root.update(tt_input)
-                return container.run()
-
-            return run_persistent
-
-        def run_inline():
-            container, _ = make_container()
-            return container.run()
-
-        return run_inline
-
-    @staticmethod
-    def _sharded_chain_unfused_callable(setup):
-        _, sharded_mem, program_cfg, tt_input, tt_w, _, _ = setup
-
-        def run_unfused():
-            result = ttnn.rms_norm(
-                tt_input,
-                weight=tt_w,
-                epsilon=1e-5,
-                program_config=program_cfg,
-                compute_kernel_config=COMPUTE_CONFIG,
-                memory_config=sharded_mem,
-            )
-            return ttnn.layer_norm(
-                result,
-                weight=tt_w,
-                epsilon=1e-5,
-                program_config=program_cfg,
-                compute_kernel_config=COMPUTE_CONFIG,
-                memory_config=sharded_mem,
-            )
-
-        return run_unfused
-
-    def _parallel_chains_fused_callable(self, setup, *, persistent):
-        from models.experimental.ops.descriptors.fusion import Parallel, Sequential
-        from models.experimental.ops.descriptors.matmul import matmul as matmul_desc
-        from models.experimental.ops.descriptors.normalization import layer_norm, rms_norm
-
-        (
-            cores_a,
-            cores_b,
-            sharded_in_a,
-            sharded_in_b,
-            sharded_out_a,
-            sharded_out_b,
-            mm_cfg_a,
-            mm_cfg_b,
-            ta,
-            tb,
-            tw,
-            tbi,
-            tB,
-        ) = setup
-
-        def make_container():
-            common_a = dict(
-                core_range_set=cores_a,
-                weight=tw,
-                epsilon=1e-5,
-                compute_kernel_config=COMPUTE_CONFIG,
-                memory_config=sharded_in_a,
-            )
-            common_b = dict(
-                core_range_set=cores_b,
-                weight=tw,
-                epsilon=1e-5,
-                compute_kernel_config=COMPUTE_CONFIG,
-                memory_config=sharded_in_b,
-            )
-            la = (
-                layer_norm.layer_norm(bias=tbi, **common_a)
-                if persistent
-                else layer_norm.layer_norm(ta, bias=tbi, **common_a)
-            )
-            rb = rms_norm.rms_norm(**common_b) if persistent else rms_norm.rms_norm(tb, **common_b)
-            ma = matmul_desc(
-                la.output_tensors[0],
-                tB,
-                core_range_set=cores_a,
-                program_config=mm_cfg_a,
-                compute_kernel_config=COMPUTE_CONFIG,
-                output_mem_config=sharded_out_a,
-            )
-            mb = matmul_desc(
-                rb.output_tensors[0],
-                tB,
-                core_range_set=cores_b,
-                program_config=mm_cfg_b,
-                compute_kernel_config=COMPUTE_CONFIG,
-                output_mem_config=sharded_out_b,
-            )
-            container = Parallel(
-                chain_a=Sequential(ln_a=la, mm_a=ma),
-                chain_b=Sequential(rms_b=rb, mm_b=mb),
-            )
-            return container, la, rb
-
-        if persistent:
-            container, la, rb = make_container()
-
-            def run_persistent():
-                la.update(ta)
-                rb.update(tb)
-                return container.run()
-
-            return run_persistent
-
-        def run_inline():
-            container, _, _ = make_container()
-            return container.run()
-
-        return run_inline
-
-    def _parallel_chains_unfused_callable(self, device, setup):
-        (
-            cores,
-            _,
-            sharded_in,
-            _,
-            sharded_out,
-            _,
-            mm_cfg,
-            _,
-            ta,
-            tb,
-            tw,
-            tbi,
-            tB,
-        ) = setup
-
-        # The unfused TTNN path requires a (0,0)-origin grid. Copy branch B's
-        # activation to the same memory layout outside the timed region.
-        tb_origin = ttnn.from_torch(
-            ttnn.to_torch(tb),
-            dtype=ttnn.bfloat16,
-            layout=ttnn.TILE_LAYOUT,
-            device=device,
-            memory_config=sharded_in,
-        )
-        ln_cfg = ttnn.LayerNormShardedMultiCoreProgramConfig(
-            compute_with_storage_grid_size=(1, 8),
-            subblock_w=4,
-            block_h=4,
-            block_w=8,
-            inplace=False,
-        )
-
-        def run_unfused():
-            ua1 = ttnn.layer_norm(
-                ta,
-                weight=tw,
-                bias=tbi,
-                epsilon=1e-5,
-                program_config=ln_cfg,
-                compute_kernel_config=COMPUTE_CONFIG,
-                memory_config=sharded_in,
-            )
-            ua2 = ttnn.matmul(
-                ua1,
-                tB,
-                program_config=mm_cfg,
-                compute_kernel_config=COMPUTE_CONFIG,
-                memory_config=sharded_out,
-            )
-            ub1 = ttnn.rms_norm(
-                tb_origin,
-                weight=tw,
-                epsilon=1e-5,
-                program_config=ln_cfg,
-                compute_kernel_config=COMPUTE_CONFIG,
-                memory_config=sharded_in,
-            )
-            ub2 = ttnn.matmul(
-                ub1,
-                tB,
-                program_config=mm_cfg,
-                compute_kernel_config=COMPUTE_CONFIG,
-                memory_config=sharded_out,
-            )
-            return ua2, ub2
-
-        return run_unfused
-
-    def _sharded_tree_fused_callable(self, device, setup, *, persistent):
-        if not persistent:
-
-            def run_inline():
-                built = self._sharded_tree_make_ops(device, setup=setup)
-                return self._sharded_tree_container(built[:13]).run()
-
-            return run_inline
-
-        built = self._sharded_tree_make_ops(device, setup=setup, persistent=True)
-        root = built[0]
-        tree = self._sharded_tree_container(built[:13])
-        tt_input = setup[10]
-
-        def run_persistent():
-            root.update(tt_input)
-            return tree.run()
-
-        return run_persistent
-
-    def _sharded_tree_unfused_callable(self, setup):
-        (
-            stem_cores,
-            left_cores,
-            _,
-            ll_cores,
-            _,
-            _,
-            _,
-            mm_cfg,
-            _,
-            mm_n,
-            tt_input,
-            tt_B_left,
-            tt_B_right,
-            shards,
-        ) = setup
-        rows, cols = 2048, 256
-        half, quarter = rows // 2, rows // 4
-
-        def _shard_mem(cores, shard_h, shard_w):
-            spec = ttnn.ShardSpec(cores, [shard_h, shard_w], ttnn.ShardOrientation.ROW_MAJOR)
-            return ttnn.MemoryConfig(ttnn.TensorMemoryLayout.BLOCK_SHARDED, ttnn.BufferType.L1, spec)
-
-        branch_mem = _shard_mem(left_cores, 128, 256)
-        mm_mem = _shard_mem(left_cores, 128, 128)
-        leaf_mem = _shard_mem(ll_cores, 128, 128)
-        stem_ln_cfg = ttnn.LayerNormShardedMultiCoreProgramConfig(
-            compute_with_storage_grid_size=(2, 8),
-            subblock_w=4,
-            block_h=8,
-            block_w=4,
-            inplace=False,
-        )
-        leaf_ln_cfg = ttnn.LayerNormShardedMultiCoreProgramConfig(
-            compute_with_storage_grid_size=(1, 4),
-            subblock_w=4,
-            block_h=4,
-            block_w=4,
-            inplace=False,
-        )
-
-        def run_unfused():
-            u_stem = ttnn.layer_norm(
-                tt_input,
-                epsilon=1e-5,
-                compute_kernel_config=COMPUTE_CONFIG,
-                program_config=stem_ln_cfg,
-                memory_config=shards["stem"],
-            )
-            u_top = ttnn.slice(u_stem, [0, 0, 0, 0], [1, 1, half, cols], memory_config=branch_mem)
-            u_bot = ttnn.slice(u_stem, [0, 0, half, 0], [1, 1, rows, cols], memory_config=branch_mem)
-            u_left = ttnn.matmul(
-                u_top,
-                tt_B_left,
-                program_config=mm_cfg,
-                compute_kernel_config=COMPUTE_CONFIG,
-                memory_config=mm_mem,
-            )
-            u_right = ttnn.matmul(
-                u_bot,
-                tt_B_right,
-                program_config=mm_cfg,
-                compute_kernel_config=COMPUTE_CONFIG,
-                memory_config=mm_mem,
-            )
-            leaves = (
-                ttnn.slice(u_left, [0, 0, 0, 0], [1, 1, quarter, mm_n], memory_config=leaf_mem),
-                ttnn.slice(u_left, [0, 0, quarter, 0], [1, 1, half, mm_n], memory_config=leaf_mem),
-                ttnn.slice(u_right, [0, 0, 0, 0], [1, 1, quarter, mm_n], memory_config=leaf_mem),
-                ttnn.slice(u_right, [0, 0, quarter, 0], [1, 1, half, mm_n], memory_config=leaf_mem),
-            )
-            return tuple(
-                ttnn.layer_norm(
-                    leaf,
-                    epsilon=1e-5,
-                    compute_kernel_config=COMPUTE_CONFIG,
-                    program_config=leaf_ln_cfg,
-                    memory_config=leaf_mem,
-                )
-                for leaf in leaves
-            )
-
-        return run_unfused
-
-    @SLIDE_PERF_SKIP
-    @pytest.mark.parametrize("mode", ["inline", "persistent", "unfused"])
-    @pytest.mark.parametrize("H", [128, 1536], ids=["H128", "H1536"])
-    def test_slide_linear_chain_e2e(self, device, H, mode):
-        setup = self._linear_chain_perf_setup(device, H)
-        mode_callables = {
-            "inline": lambda: self._linear_chain_fused_callable(setup, persistent=False),
-            "persistent": lambda: self._linear_chain_fused_callable(setup, persistent=True),
-            "unfused": lambda: self._linear_chain_unfused_callable(setup),
-        }
-        median_us, samples_us = _benchmark_e2e_mode(mode_callables[mode](), device)
-        _print_slide_perf_result(f"linear_chain_H{H}", mode, median_us, samples_us)
-
-    @SLIDE_PERF_SKIP
-    @pytest.mark.parametrize("mode", ["inline", "persistent", "unfused"])
-    @pytest.mark.parametrize("H", [128, 1536], ids=["H128", "H1536"])
-    def test_slide_sharded_chain_e2e(self, device, H, mode):
-        setup = self._sharded_chain_setup(device, H)
-        mode_callables = {
-            "inline": lambda: self._sharded_chain_fused_callable(setup, persistent=False),
-            "persistent": lambda: self._sharded_chain_fused_callable(setup, persistent=True),
-            "unfused": lambda: self._sharded_chain_unfused_callable(setup),
-        }
-        median_us, samples_us = _benchmark_e2e_mode(mode_callables[mode](), device)
-        _print_slide_perf_result(f"sharded_chain_H{H}", mode, median_us, samples_us)
-
-    @SLIDE_PERF_SKIP
-    @pytest.mark.parametrize("mode", ["inline", "persistent", "unfused"])
-    def test_slide_parallel_chains_e2e(self, device, mode):
-        setup = self._parallel_chains_setup(device)
-        mode_callables = {
-            "inline": lambda: self._parallel_chains_fused_callable(setup, persistent=False),
-            "persistent": lambda: self._parallel_chains_fused_callable(setup, persistent=True),
-            "unfused": lambda: self._parallel_chains_unfused_callable(device, setup),
-        }
-        median_us, samples_us = _benchmark_e2e_mode(mode_callables[mode](), device)
-        _print_slide_perf_result("parallel_chains", mode, median_us, samples_us)
-
-    def test_parallel_chains_persistent_releases_semaphore_bank_l1(self, device):
-        run = self._parallel_chains_fused_callable(self._parallel_chains_setup(device), persistent=True)
-        for _ in range(5):
-            run()
-        ttnn.synchronize_device(device)
-        gc.collect()
-        pages_before = len(ttnn._ttnn.reports.get_buffer_pages(device))
-
-        for _ in range(100):
-            run()
-        ttnn.synchronize_device(device)
-        gc.collect()
-
-        assert len(ttnn._ttnn.reports.get_buffer_pages(device)) == pages_before
-
-    @SLIDE_PERF_SKIP
-    @pytest.mark.parametrize("mode", ["inline", "persistent", "unfused"])
-    def test_slide_balanced_tree_e2e(self, device, mode):
-        setup = self._sharded_tree_setup(device)
-
-        mode_callables = {
-            "inline": lambda: self._sharded_tree_fused_callable(device, setup, persistent=False),
-            "persistent": lambda: self._sharded_tree_fused_callable(device, setup, persistent=True),
-            "unfused": lambda: self._sharded_tree_unfused_callable(setup),
-        }
-        median_us, samples_us = _benchmark_e2e_mode(mode_callables[mode](), device)
-        _print_slide_perf_result("balanced_tree", mode, median_us, samples_us)
-
-    @SLIDE_PERF_SKIP
-    @pytest.mark.parametrize("mode", ["inline", "persistent", "unfused"])
-    @pytest.mark.parametrize("H", [128, 1536], ids=["H128", "H1536"])
-    def test_slide_linear_chain_device_fw(self, device, H, mode):
-        """Repeated-dispatch entry point for manual device profiling."""
-        setup = self._linear_chain_perf_setup(device, H)
-        mode_callables = {
-            "inline": lambda: self._linear_chain_fused_callable(setup, persistent=False),
-            "persistent": lambda: self._linear_chain_fused_callable(setup, persistent=True),
-            "unfused": lambda: self._linear_chain_unfused_callable(setup),
-        }
-        run = mode_callables[mode]()
-        _run_device_profile(run, device)
-
-    @SLIDE_PERF_SKIP
-    @pytest.mark.parametrize("mode", ["inline", "persistent", "unfused"])
-    @pytest.mark.parametrize("H", [128, 1536], ids=["H128", "H1536"])
-    def test_slide_sharded_chain_device_fw(self, device, H, mode):
-        """Repeated-dispatch entry point for manual device profiling."""
-        setup = self._sharded_chain_setup(device, H)
-        mode_callables = {
-            "inline": lambda: self._sharded_chain_fused_callable(setup, persistent=False),
-            "persistent": lambda: self._sharded_chain_fused_callable(setup, persistent=True),
-            "unfused": lambda: self._sharded_chain_unfused_callable(setup),
-        }
-        run = mode_callables[mode]()
-        _run_device_profile(run, device)
-
-    @SLIDE_PERF_SKIP
-    @pytest.mark.parametrize("mode", ["inline", "persistent"])
-    def test_slide_parallel_chains_fused_device_fw(self, device, mode):
-        """Repeated-dispatch entry point for manual device profiling."""
-        setup = self._parallel_chains_setup(device)
-        run = self._parallel_chains_fused_callable(setup, persistent=mode == "persistent")
-        _run_device_profile(run, device)
-
-    @SLIDE_PERF_SKIP
-    def test_slide_parallel_chains_unfused_device_fw(self, device):
-        """Repeated-dispatch entry point for manual device profiling."""
-        setup = self._parallel_chains_setup(device)
-        _run_device_profile(self._parallel_chains_unfused_callable(device, setup), device)
-
-    @SLIDE_PERF_SKIP
-    @pytest.mark.parametrize("mode", ["inline", "persistent"])
-    def test_slide_balanced_tree_fused_device_fw(self, device, mode):
-        """Repeated-dispatch entry point for manual device profiling."""
-        setup = self._sharded_tree_setup(device)
-        run = self._sharded_tree_fused_callable(device, setup, persistent=mode == "persistent")
-        _run_device_profile(run, device)
-
-    @SLIDE_PERF_SKIP
-    def test_slide_balanced_tree_unfused_device_fw(self, device):
-        """Repeated-dispatch entry point for manual device profiling."""
-        setup = self._sharded_tree_setup(device)
-        _run_device_profile(self._sharded_tree_unfused_callable(setup), device)
-
 
 # =============================================================================
 # Functional Demos (fused only, no unfused comparison)
@@ -3019,28 +2428,3 @@ def test_barrier_overhead(device, num_phases, num_cores, perf_mode):
         per_barrier_us = (fused_us - baseline_us) / (num_phases - 1)
     else:
         pytest.fail(f"Unsupported perf_mode={perf_mode!r}")
-
-
-@SLIDE_PERF_SKIP
-@pytest.mark.skipif(
-    is_watcher_enabled(), reason="pytest-timeout plugin interacts with watcher on device reopen (noop kernels)"
-)
-@pytest.mark.parametrize("num_cores", [1, 8, 16, 64])
-@pytest.mark.parametrize("num_phases", [2, 3, 4, 5, 6])
-def test_slide_barrier_overhead_e2e(device, num_phases, num_cores):
-    """Measure median per-transition barrier cost for the documentation table."""
-    from models.experimental.ops.descriptors.fusion import Sequential
-
-    fused = Sequential(*_barrier_bench_setup(device, num_phases, num_cores))
-    baseline = Sequential(*_barrier_bench_setup(device, 1, num_cores))
-    median_us, samples_us = _benchmark_barrier_per_transition(
-        fused.run,
-        baseline.run,
-        device,
-        num_transitions=num_phases - 1,
-    )
-    samples = ",".join(f"{sample:.3f}" for sample in samples_us)
-    print(
-        f"\nFUSION_BARRIER_PERF cores={num_cores} phases={num_phases} "
-        f"median_us={median_us:.3f} samples_us={samples}"
-    )
