+++ tests/ttnn/unit_tests/operations/fused/parallel_sequential/test_parallel_sequential_infra.py
@@ -5,12 +5,17 @@
 """
 Standalone tests for Sequential Kernel Chaining Infrastructure.
 
-Tests the core fusion infrastructure logic without requiring a device.
+Most tests exercise fusion logic without a device. Semaphore-bank tests
+require hardware.
 """
 
-import pytest
+import gc
+import time
 from types import SimpleNamespace
 
+import pytest
+import torch
+import ttnn
 
 import models.experimental.ops.descriptors.fusion.common as _common
 import models.experimental.ops.descriptors.fusion.cb_allocator as _cb_alloc
@@ -1293,6 +1298,117 @@ class TestCompileTimePerformance:
         assert t8 < 10.0, f"8-phase took {t8:.2f}s"
         if t2 > 0.001:
             assert t8 / t2 < 20.0, f"Ratio {t8/t2:.1f}x exceeds 20x"
+
+
+# ---------------------------------------------------------------------------
+# Fusion semaphore bank
+# ---------------------------------------------------------------------------
+
+
+def _semaphore_core(x, y):
+    coord = ttnn.CoreCoord(x, y)
+    return ttnn.CoreRangeSet({ttnn.CoreRange(coord, coord)})
+
+
+def _semaphore_cores(*coords):
+    return ttnn.CoreRangeSet({ttnn.CoreRange(ttnn.CoreCoord(x, y), ttnn.CoreCoord(x, y)) for x, y in coords})
+
+
+class TestFusionSemaphoreBank:
+    def test_nonzero_values_and_offsets(self, device):
+        core_ranges = [_semaphore_core(0, 0), _semaphore_core(2, 0), _semaphore_cores((0, 0), (2, 0))]
+        initial_values = [3, 7, 11]
+
+        bank = ttnn._ttnn.operations.experimental.FusionSemaphoreBank(device, core_ranges, initial_values)
+
+        base_address = bank.tensor.buffer_address()
+        assert not bank.tensor.is_per_core_allocated()
+        assert list(bank.addresses) == [base_address + 4 * index for index in range(len(initial_values))]
+
+        values = ttnn.to_torch(ttnn.from_device(bank.tensor)).reshape(-1, len(initial_values))
+        expected = torch.tensor(initial_values, dtype=torch.uint32).repeat(values.shape[0], 1)
+        assert torch.equal(values, expected)
+
+    def test_releases_l1(self, device):
+        core_ranges = [_semaphore_core(0, 0), _semaphore_core(1, 0)]
+        ttnn.synchronize_device(device)
+        gc.collect()
+        pages_before = len(ttnn._ttnn.reports.get_buffer_pages(device))
+
+        for _ in range(100):
+            bank = ttnn._ttnn.operations.experimental.FusionSemaphoreBank(device, core_ranges, [0, 0])
+            assert len(bank.addresses) == 2
+        del bank
+        ttnn.synchronize_device(device)
+        gc.collect()
+
+        assert len(ttnn._ttnn.reports.get_buffer_pages(device)) == pages_before
+
+    def test_rejects_mismatched_metadata(self, device, expect_error):
+        with expect_error(RuntimeError, "must match"):
+            ttnn._ttnn.operations.experimental.FusionSemaphoreBank(device, [_semaphore_core(0, 0)], [0, 1])
+
+    def test_rejects_empty_metadata(self, device, expect_error):
+        with expect_error(RuntimeError, "at least one semaphore"):
+            ttnn._ttnn.operations.experimental.FusionSemaphoreBank(device, [], [])
+
+    def test_mesh_address_contract(self, mesh_device):
+        bank = ttnn._ttnn.operations.experimental.FusionSemaphoreBank(
+            mesh_device,
+            [_semaphore_cores((0, 0), (1, 0)), _semaphore_core(1, 0)],
+            [0, 9],
+        )
+
+        assert not bank.tensor.is_per_core_allocated()
+        assert len(bank.tensor.device_coords()) == mesh_device.get_num_devices()
+        assert list(bank.addresses) == [bank.tensor.buffer_address(), bank.tensor.buffer_address() + 4]
+
+    def test_persistent_dispatch_releases_l1(self, device):
+        from models.experimental.ops.descriptors.fusion import Sequential
+        from models.experimental.ops.descriptors.normalization import rms_norm
+
+        core_range = ttnn.CoreRangeSet({ttnn.CoreRange(ttnn.CoreCoord(0, 0), ttnn.CoreCoord(0, 0))})
+        compute_cfg = ttnn.WormholeComputeKernelConfig(
+            math_fidelity=ttnn.MathFidelity.HiFi4,
+            math_approx_mode=False,
+            fp32_dest_acc_en=False,
+        )
+        dram = ttnn.DRAM_MEMORY_CONFIG
+        activation = ttnn.from_torch(
+            torch.randn(1, 1, 32, 32, dtype=torch.bfloat16),
+            dtype=ttnn.bfloat16,
+            layout=ttnn.TILE_LAYOUT,
+            device=device,
+            memory_config=dram,
+        )
+        weight = ttnn.from_torch(
+            torch.ones(1, 1, 1, 32, dtype=torch.bfloat16),
+            dtype=ttnn.bfloat16,
+            layout=ttnn.TILE_LAYOUT,
+            device=device,
+            memory_config=dram,
+        )
+        kwargs = dict(core_range_set=core_range, weight=weight, epsilon=1e-5, compute_kernel_config=compute_cfg)
+        rms1 = rms_norm.rms_norm(**kwargs)
+        rms2 = rms_norm.rms_norm(rms1.output_tensors[0], **kwargs)
+        container = Sequential(rms1, rms2)
+
+        def run():
+            rms1.update(activation)
+            return container.run()
+
+        for _ in range(5):
+            run()
+        ttnn.synchronize_device(device)
+        gc.collect()
+        pages_before = len(ttnn._ttnn.reports.get_buffer_pages(device))
+
+        for _ in range(100):
+            run()
+        ttnn.synchronize_device(device)
+        gc.collect()
+
+        assert len(ttnn._ttnn.reports.get_buffer_pages(device)) == pages_before
 
 
 if __name__ == "__main__":
